+++ packages/werkbank/lib/src/_internal/src/widgets/src/pages/main_page.dart
@@ -20,20 +20,18 @@ class MainPage extends StatelessWidget {
     return NavStateProvider(
       child: Historiographer(
         child: WerkbankShortcuts(
-          child: AcknowledgedTracker(
-            child: Scaffold(
-              backgroundColor: PageBackground.colorOf(context),
-              body: AddonLayerBuilder(
-                layer: AddonLayer.applicationOverlay,
-                child: UseCaseControllerManager(
-                  child: WResizablePanels(
-                    leftPanel: const NavigationPanel(),
-                    rightPanel: const ConfigurationPanel(),
-                    child: WerkbankNotificationsDisplay(
-                      child: AddonLayerBuilder(
-                        layer: AddonLayer.mainViewOverlay,
-                        child: mainView,
-                      ),
+          child: Scaffold(
+            backgroundColor: PageBackground.colorOf(context),
+            body: AddonLayerBuilder(
+              layer: AddonLayer.applicationOverlay,
+              child: UseCaseControllerManager(
+                child: WResizablePanels(
+                  leftPanel: const NavigationPanel(),
+                  rightPanel: const ConfigurationPanel(),
+                  child: WerkbankNotificationsDisplay(
+                    child: AddonLayerBuilder(
+                      layer: AddonLayer.mainViewOverlay,
+                      child: mainView,
                     ),
                   ),
                 ),