+++ packages/werkbank/lib/src/_internal/src/widgets/src/panels/configuration_panel/tab_addon_control_section_list.dart
@@ -83,7 +83,7 @@ class _AddonControlSectionListState extends State<_AddonControlSectionList> {
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    _panelTabsController = GlobalStateManager.maybePanelTabsController(
+    _panelTabsController = GlobalStateManager.maybePanelTabsControllerOf(
       context,
     )!;
     if (_orderedSections == null) {