+++ packages/werkbank/lib/src/global_state/global_state.dart
@@ -3,10 +3,6 @@
 /// Generated using index_generator package.
 library;
 
-export 'src/acknowledged/acknowledged_controller.dart';
-export 'src/acknowledged/acknowledged_controller_impl.dart';
-export 'src/acknowledged/acknowledged_descriptors.dart';
-export 'src/acknowledged/acknowledged_tracker.dart';
 export 'src/global_state_config.dart';
 export 'src/global_state_controller.dart';
 export 'src/global_state_controller_registry.dart';