+++ packages/werkbank/lib/src/global_state/src/acknowledged/acknowledged_controller.dart
@@ -1,9 +1,0 @@
-import 'package:werkbank/src/global_state/global_state.dart';
-
-abstract class AcknowledgedController implements GlobalStateController {
-  AcknowledgedDescriptors get descriptors;
-
-  void log(String path);
-
-  // Clearing is more like resetting for this feature.
-}