+++ packages/werkbank/lib/src/global_state/src/history/historiographer.dart
@@ -3,6 +3,7 @@ import 'package:werkbank/src/_internal/src/routing/routing.dart';
 import 'package:werkbank/src/_internal/src/widgets/widgets.dart';
 import 'package:werkbank/src/routing/routing.dart';
 
+// TODO: Move to somewhere internal.
 class Historiographer extends StatefulWidget {
   const Historiographer({
     required this.child,